requirements.txt
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
marshmallow
flask_marshmallow
marshmallow-sqlalchemy
pydrive2
gunicorn 
google.cloud.logging
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1